First Big Commit, Read Desc
Now generates A list of lists of files in found in the main Assets folder sorted by their parent folders, then generates a list of used assets and their character codes based off of a txt file key, for info on formatting, look at the REALREADME.txt.

Right now all the program does is print the names of the assets you will be using in standard output. Also keep in mind there is no error checking, so make sure your names are correct in the Key. And example key and some assets from SSRB are included.